Move TestMain into its own file
The testenv and test HTTP server are (or will be) used by tests in
eval_test.go and generate_test.go, and it makes sense to have it
separate from both of those.
