+++ internal/eval/main_test.go
@@ -1,0 +1,52 @@
+/*
+Copyright 2023 Michael Bridgen.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package eval
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/envtest"
+)
+
+var (
+	k8sClient client.Client
+	baseurl   string
+)
+
+func TestMain(m *testing.M) {
+	testEnv := &envtest.Environment{}
+	cfg, err := testEnv.Start()
+	if err != nil {
+		panic(err)
+	}
+	defer testEnv.Stop()
+
+	c, err := client.New(cfg, client.Options{})
+	if err != nil {
+		panic(err)
+	}
+	k8sClient = c
+
+	httpserver := httptest.NewServer(http.FileServer(http.Dir("testdata")))
+	baseurl = httpserver.URL
+	defer httpserver.Close()
+
+	m.Run()
+}
